Extract common code to be able to start changing
