+++ dockers/llm.rag.service/Dockerfile
@@ -18,7 +18,9 @@ RUN apt-get update && DEBIAN_FRONTEND=noninteractive apt-get install -y --no-ins
 
 WORKDIR /serveragllm
 
+COPY __init__.py .
 COPY serveragllm.py .
+COPY serverragllm_jira_cvs_local.py .
 COPY pyproject.toml .
 
 RUN --mount=type=cache,target=/root/.cache/pip \
