+++ dockers/llm.rag.service/serveragllm.py
@@ -8,6 +8,8 @@ from botocore.exceptions import ClientError, NoCredentialsError
 from fastapi import FastAPI
 from openai import OpenAI
 
+from common import get_answer_with_settings
+
 ########
 # Setup model name and query template parameters
 MICROSOFT_MODEL_ID = "microsoft/Phi-3-mini-4k-instruct"
@@ -47,24 +49,6 @@ def str_to_float(value, name):
         )
         sys.exit(1)
     return float_value
-
-
-def format_context(results: List[Dict[str, Any]]) -> str:
-    """Format search results into context for the LLM"""
-    context_parts = []
-
-    for result in results:
-        # TODO: make metadata keyes configurable
-        ticket_metadata = result.metadata
-        ticket_content = result.page_content
-
-        context_parts.append(
-            f"Key: {ticket_metadata['key']} | Status: {ticket_metadata['status']} - "
-            f"Type: {ticket_metadata['type']}\n"
-            f"Content: {ticket_content}...\n"
-        )
-
-    return "\n\n".join(context_parts)
 
 
 ########
@@ -109,38 +93,14 @@ def get_answer(question: Union[str, None]):
     )
 
     if is_jira_mode:
-        SYSTEM_PROMPT = """You are a specialized Jira ticket assistant. Format your responses following these rules:
-            1. Start with the most relevant ticket references
-            2. Provide a clear, direct answer
-            3. Include relevant technical details when present
-            4. Mention related tickets if they provide additional context
-            5. If the information is outdated, mention when it was last updated
-            """
-
-        context = format_context(docs)
-
-        print("Sending query to the LLM...")
-        completions = client.chat.completions.create(
-            model=model_id,
-            messages=[
-                {"role": "system", "content": SYSTEM_PROMPT},
-                {
-                    "role": "user",
-                    "content": f"Context:\n{context}\n\nQuestion: {question}",
-                },
-            ],
-            max_tokens=max_tokens,
-            temperature=model_temperature,
-            stream=False,
+        return get_answer_with_settings(
+            question,
+            retriever,
+            client,
+            model_id,
+            max_tokens,
+            model_temperature,
         )
-
-        answer = {
-            "answer": completions.choices[0].message.content,
-            "relevant_tickets": [r.metadata["key"] for r in docs],
-        }
-        print("Received answer: ", answer)
-        return answer
-
     else:
         print("Sending query to the LLM...")
         # concatenate relevant docs retrieved to be used as context
